Changed `match-repository` to not use points
Functions previously stored the number of points scored for each task.
Now, they just store the number of completions of that task. Added
functions for getting and setting the phase and the number of power
cells scored in the most recent phase. Separated hang and park points.
Added function for checking if the rung was level.